#210 [test] searchUseCase
